welp

+++ Text Diffusion/src/utils.py
@@ -13,7 +13,7 @@ def log_add(log_x, log_y):
     return maximum + torch.log(torch.exp(log_x - maximum) + torch.exp(log_y - maximum))
 
 def index_to_onehot(x, num_classes, use_log=False):
-    x_onehot = F.one_hot(x, num_classes)
+    x_onehot = F.one_hot(x, num_classes).float()
 
     if use_log:
         x_onehot = torch.log(x_onehot.float().clamp(min=1e-30))
